requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-filter==22.1
django-rest==0.8.7
djangorestframework==3.14.0
drf-extra-fields==3.4.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2==2.9.5
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
